first draft of more join operations exercises and database done
took in rithm's existing movies database and added one column to the roles table, lead_role, to be a boolean value in an attempt to match the sqlzoo version, somewhat anyway. set the default value of that to false and updated about two dozen entries in the roles table (based on actors likely to be in leading roles) to hold a lead_role value of true. now, this can also be removed, but a good portion of the exercises from sqlzoo rely on this idea and i wanted to play around with it. will probably take another look tomorrow but wanted to push this up for now.