Add name and join exercises to zoo (#1)
* initial commit, added select-names yaml file

in the process of removing the mysql stuff out of the euro sql file, just realized i hadn't committed yet

* euro data formatted to postgres, loads to zoo_local successfully

* nearly finished with first set of join exercises

renamed game table to be called match instead, for consistency's sake. otherwise, working through the rest of the join exercises.

* nearly done with join exercises, including those labeled as more difficult

committing now as there are some changes made to the main branch that i'd like to merge in. still working on this branch and will look for some tidying opportunities once i finish adding the exercises

* join exercises finished, including those labeled more difficult

have also added all but one of the more-difficult questions for the names module. the remaining question is worded strangely on sqlzoo, so it isn't completely clear to me what they're asking, at least right now. still need to add one more table description into the schema section of the join exercise. also likely have some format tidying to do before adding in the last more_join_operations module

* first draft of more join operations exercises and database done

took in rithm's existing movies database and added one column to the roles table, lead_role, to be a boolean value in an attempt to match the sqlzoo version, somewhat anyway. set the default value of that to false and updated about two dozen entries in the roles table (based on actors likely to be in leading roles) to hold a lead_role value of true. now, this can also be removed, but a good portion of the exercises from sqlzoo rely on this idea and i wanted to play around with it. will probably take another look tomorrow but wanted to push this up for now.

* tidying the names module a bit

just some small formatting changes. also, may need to improve the instructions for the very last exercise of the select names module. i'm fairly sure what i've provided is what they're asking, but didn't get the success message on sqlzoo for this one. may need a second opinion.

* tidying the join exercise module a bit

added in table names to some solutions where i didn't think it was completely obvious where the column was coming from, mostly in the join lines themselves. still some formatting weirdness on the example table in the last question description but maybe that isn't imperative

* small typo in join module

was trying a few things with the pipe operator since i think the formatting issue is related to that in some way-- forgot to remove it so that's all this commit is

* a little tidying and some typo fixes

nothing major here, just some last reviewing

* removed extra column from movies table

original schema included a column for studio_id in the movies table, which isn't a concept that appears in the SQLZoo movies database. removed from the movies.sql file and removed that value insertion from all movie records.

* Tidy formatting and SQL idioms.

- SQL keywords should always be ALL CAPS

- CONCAT(x, y) is a gross MySQL-ism and is not in the SQL
  spec; changed to `x || y`

- Switched some LIKE -> ILIKE where it makes sense

Co-authored-by: Joel Burton <joel@joelburton.com>